+++ static/css/upload.css
@@ -1,0 +1,25 @@
+* {
+    margin: 0;
+    padding: 0;
+}
+
+.upload-container {
+    position: absolute;
+    left: 50%;
+    top: 50%;
+    transform:translate(-50%,-50%);
+}
+
+.upload-container textarea {
+    display: inline-block;
+    border: 1px solid #000;
+}
+
+.upload-container button {
+    display: block;
+    margin: 0 auto;
+    width: 30%;
+    height: 30px;
+    border-radius: 4px;
+    background: #fff;
+}
